Декомпозирует редьюсер, добавляет селекторы
