+++ src/components/app/app.jsx
@@ -3,12 +3,14 @@ import {connect} from "react-redux";
 import React, {PureComponent} from "react";
 import PropTypes from "prop-types";
 
-import {ActionCreator} from "../../reducer/reducer.js";
+import {ActionCreator} from "../../reducer/game/game.js";
 import ArtistQuestionScreen from "../artist-question-screen/artist-question-screen.jsx";
 import GameOverScreen from "../game-over-screen/game-over-screen.jsx";
 import GameScreen from "../game-screen/game-screen.jsx";
 import {GameType, START_STEP} from "../../const.js";
 import GenreQuestionScreen from "../genre-question-screen/genre-question-screen.jsx";
+import {getMaxMistakes, getMistakes, getStep} from "../../reducer/game/selectors.js";
+import {getQuestions} from "../../reducer/data/selectors.js";
 import WelcomeScreen from "../welcome-screen/welcome-screen.jsx";
 import WinScreen from "../win-screen/win-screen.jsx";
 import withActivePlayer from "../../hocs/with-active-player/with-active-player.js";
@@ -121,10 +123,10 @@ App.propTypes = {
 };
 
 const mapStateToProps = (state) => ({
-  mistakes: state.mistakes,
-  maxMistakes: state.maxMistakes,
-  step: state.step,
-  questions: state.questions,
+  mistakes: getMistakes(state),
+  maxMistakes: getMaxMistakes(state),
+  step: getStep(state),
+  questions: getQuestions(state),
 });
 
 const mapDispatchToProps = (dispatch) => ({
