+++ src/components/game-screen/game-screen.jsx
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import React from "react";
 
 import {GameType} from "../../const.js";
-
+import {getMistakes} from "../../reducer/game/selectors.js";
 import Mistakes from "../mistakes/mistakes.jsx";
 
 const GameScreen = (props) => {
@@ -42,7 +42,7 @@ GameScreen.propTypes = {
 };
 
 const mapStateToProps = (state) => ({
-  mistakes: state.mistakes,
+  mistakes: getMistakes(state),
 });
 
 export {GameScreen};
