+++ src/index.js
@@ -7,7 +7,8 @@ import thunk from "redux-thunk";
 
 import App from "./components/app/app.jsx";
 import createAPI from "./api.js";
-import {reducer, Operation as DataOperation} from "./reducer/reducer.js";
+import {Operation as DataOperation} from "./reducer/data/data.js";
+import reducer from "./reducer/reducer.js";
 
 const api = createAPI(() => {});
 const store = createStore(
