+++ src/reducer/data/selectors.js
@@ -1,0 +1,5 @@
+import NameSpace from "../name-space.js";
+
+const NAME_SPACE = NameSpace.DATA;
+
+export const getQuestions = (state) => state[NAME_SPACE].questions;
