+++ src/reducer/game/game.test.js
@@ -1,67 +1,16 @@
-import {ActionCreator, ActionType, Operation, reducer} from "./reducer.js";
-import createAPI from "../api.js";
-import MockAdapter from "axios-mock-adapter";
+import {ActionCreator, ActionType, reducer} from "./game.js";
 
 const testInitialState = {
   mistakes: 0,
   maxMistakes: 3,
   step: -1,
-  questions: []
 };
-
-const testQuestions = [{
-  type: `genre`,
-  genre: `rock`,
-  answers: [{
-    src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
-    genre: `rock`,
-  }, {
-    src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
-    genre: `blues`,
-  }, {
-    src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
-    genre: `jazz`,
-  }, {
-    src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
-    genre: `rock`,
-  }],
-}, {
-  type: `artist`,
-  song: {
-    artist: `Jim Beam`,
-    src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
-  },
-  answers: [{
-    picture: `https://api.adorable.io/avatars/128/1`,
-    artist: `John Snow`,
-  }, {
-    picture: `https://api.adorable.io/avatars/128/2`,
-    artist: `Jack Daniels`,
-  }, {
-    picture: `https://api.adorable.io/avatars/128/3`,
-    artist: `Jim Beam`,
-  }],
-}];
-
-const api = createAPI(() => {});
 
 describe(`Reducer work properly`, () => {
   it(`should return initial state without additional parameters`, () => {
     const initialReducer = reducer(undefined, {});
 
     expect(initialReducer).toEqual(testInitialState);
-  });
-
-  it(`should update questions by load`, () => {
-    expect(reducer(testInitialState, {
-      type: ActionType.LOAD_QUESTIONS,
-      payload: testQuestions,
-    })).toEqual({
-      mistakes: 0,
-      maxMistakes: 3,
-      step: -1,
-      questions: testQuestions
-    });
   });
 
   it(`should increment number of mistakes by a given value`, () => {
@@ -72,7 +21,6 @@ describe(`Reducer work properly`, () => {
       mistakes: 1,
       maxMistakes: 3,
       step: -1,
-      questions: []
     });
 
     expect(reducer(testInitialState, {
@@ -82,7 +30,6 @@ describe(`Reducer work properly`, () => {
       mistakes: 0,
       maxMistakes: 3,
       step: -1,
-      questions: []
     });
   });
 
@@ -94,7 +41,6 @@ describe(`Reducer work properly`, () => {
       mistakes: 0,
       maxMistakes: 3,
       step: 0,
-      questions: []
     });
 
     expect(reducer(testInitialState, {
@@ -104,7 +50,6 @@ describe(`Reducer work properly`, () => {
       mistakes: 0,
       maxMistakes: 3,
       step: -1,
-      questions: []
     });
   });
 
@@ -258,23 +203,3 @@ describe(`Action creators work properly`, () => {
   });
 });
 
-describe(`Operation work properly`, () => {
-  it(`make a correct API call to /questions`, () => {
-    const apiMock = new MockAdapter(api);
-    const dispatch = jest.fn();
-    const questionLoader = Operation.loadQuestions();
-
-    apiMock
-      .onGet(`/questions`)
-      .reply(200, [{fake: true}]);
-
-    return questionLoader(dispatch, () => {}, api)
-      .then(() => {
-        expect(dispatch).toHaveBeenCalledTimes(1);
-        expect(dispatch).toHaveBeenCalledWith({
-          type: ActionType.LOAD_QUESTIONS,
-          payload: [{fake: true}],
-        });
-      });
-  });
-});
