+++ src/reducer/name-space.js
@@ -1,0 +1,6 @@
+const NameSpace = {
+  DATA: `DATA`,
+  GAME: `GAME`,
+};
+
+export default NameSpace;
